Automatic patching prototype.
This is an attempt to fix https://github.com/acsone/setuptools-odoo/issues/10

By installing a .pth file named zzz_odoo_autodiscover.pth that will run last,
but before the main program starts, we can remove the dummy odoo package that
is created by the *nspkg.pth files from setuptools. That leaves a chance to
odoo/__init__.py to run when the main program imports odoo. The declare_namespace
in odoo/addons/__init__.py will then repopulate odoo.addons.__path__.

Yes, that is tricky... it's the price to live on the border line where.
To be spec compliant, odoo/__init__.py should be empty. A better solution is
to be found for v11.